tweaked the interaface search adapter (may not be finished)
