+++ src/main/Java/interface_adapter/search/SearchState.java
@@ -4,19 +4,27 @@ import entity.Event;
 import java.util.List;
 
 public class SearchState {
-    private final List<Event> events;
-    private final String errorMessage;
+    private List<Event> results;
+    private String error;
 
-    public SearchState(List<Event> events, String errorMessage) {
-        this.events = events;
-        this.errorMessage = errorMessage;
+    public SearchState(List<Event> results, String error) {
+        this.results = results;
+        this.error = error;
     }
 
-    public List<Event> getEvents() {
-        return events;
+    public List<Event> getResults() {
+        return results;
     }
 
-    public String getErrorMessage() {
-        return errorMessage;
+    public String getError() {
+        return error;
+    }
+
+    public void setResults(List<Event> results) {
+        this.results = results;
+    }
+
+    public void setError(String error) {
+        this.error = error;
     }
 }
