+++ src/main/Java/interface_adapter/search/SearchViewModel.java
@@ -1,26 +1,24 @@
 package interface_adapter.search;
 
 import entity.Event;
+import interface_adapter.ViewModel;
 
 import java.util.List;
 
-public class SearchViewModel {
-    private List<Event> results;
-    private String error;
-
-    public List<Event> getResults() {
-        return results;
+public class SearchViewModel extends ViewModel<SearchState> {
+    public SearchViewModel(String viewName) {
+        super(viewName);
     }
 
-    public void setResults(List<Event> results) {
-        this.results = results;
+    public void setSearchResults(List<Event> results) {
+        SearchState newState = new SearchState(results, null);
+        setState(newState);  // Update state with results and no error
+        firePropertyChanged();
     }
 
-    public String getError() {
-        return error;
-    }
-
-    public void setError(String error) {
-        this.error = error;
+    public void setError(String errorMessage) {
+        SearchState newState = new SearchState(null, errorMessage);
+        setState(newState);  // Update state with error
+        firePropertyChanged();
     }
 }
